+++ src/main/java/com/levelupfit/mainbackend/dto/user/FormUserDTO.java
@@ -1,19 +1,16 @@
-package com.levelupfit.mainbackend.dto;
+package com.levelupfit.mainbackend.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-
-public class SocialUserDTO {
+public class FormUserDTO {
 
     private String userId;
-    private String providerType;
-    private String email;
+    private String pwd;
 }
