+++ src/main/java/com/levelupfit/mainbackend/dto/user/SocialUserDTO.java
@@ -1,16 +1,19 @@
-package com.levelupfit.mainbackend.dto;
+package com.levelupfit.mainbackend.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class FormUserDTO {
+
+public class SocialUserDTO {
 
     private String userId;
-    private String pwd;
+    private String providerType;
+    private String email;
 }
