+++ src/main/java/com/levelupfit/mainbackend/dto/user/UpdateNicknameDTO.java
@@ -1,4 +1,4 @@
-package com.levelupfit.mainbackend.dto;
+package com.levelupfit.mainbackend.dto.user;
 
 import lombok.Data;
